refactor(web): decompose terminal log viewer
