+++ apps/web/src/components/terminal-log-viewer/terminal-log-viewer.tsx
@@ -1,0 +1,215 @@
+import { useCallback, useMemo, useRef, useState } from 'react'
+import { AlertCircleIcon } from '@hugeicons/core-free-icons'
+import { HugeiconsIcon } from '@hugeicons/react'
+import { useVirtualizer } from '@tanstack/react-virtual'
+
+import { LogOutput } from './log-output'
+import { LogToolbar } from './log-toolbar'
+import {
+  defaultSelectedStep,
+  findFirstErrorIndex,
+  groupLogs,
+} from './log-model'
+import { MobileStepSelector, StepSidebar } from './step-navigation'
+import type { SelectedStepMeta, TerminalLogViewerProps } from './types'
+import { useWindowEvent } from '@/hooks/use-window-event'
+import { useMountEffect } from '@/hooks/use-mount-effect'
+import { useAutoScroll } from '@/hooks/use-auto-scroll'
+import { Alert, AlertDescription } from '@/components/ui/alert'
+
+export default function TerminalLogViewer({
+  logs,
+  stepResults,
+  isStreaming,
+  streamError,
+  logsUnavailable = false,
+  isTerminal = false,
+}: TerminalLogViewerProps) {
+  const [userSelectedStep, setUserSelectedStep] = useState<string | null>(null)
+  const [autoScroll, setAutoScroll] = useState(true)
+  const [wrapLines, setWrapLines] = useState(false)
+  const [searchQuery, setSearchQuery] = useState('')
+  const [searchOpen, setSearchOpen] = useState(false)
+  const scrollContainerRef = useRef<HTMLDivElement>(null)
+  const searchInputRef = useRef<HTMLInputElement>(null)
+
+  const { stepGroups, stepGroupsByName, allVisibleLogs, runningStepName } =
+    useMemo(() => groupLogs(logs, stepResults), [logs, stepResults])
+
+  const selectedStep = useMemo(() => {
+    if (
+      userSelectedStep === 'all' ||
+      (userSelectedStep !== null && stepGroupsByName.has(userSelectedStep))
+    ) {
+      return userSelectedStep
+    }
+    return defaultSelectedStep(stepGroups, runningStepName)
+  }, [userSelectedStep, stepGroups, stepGroupsByName, runningStepName])
+
+  const selectedLogs = useMemo(
+    () =>
+      selectedStep === 'all'
+        ? allVisibleLogs
+        : (stepGroupsByName.get(selectedStep)?.logs ?? []),
+    [selectedStep, allVisibleLogs, stepGroupsByName],
+  )
+  const filteredLogs = useMemo(() => {
+    if (!searchQuery.trim()) return selectedLogs
+    const query = searchQuery.toLowerCase()
+    return selectedLogs.filter((chunk) =>
+      chunk.content.toLowerCase().includes(query),
+    )
+  }, [selectedLogs, searchQuery])
+  const selectedStepMeta: SelectedStepMeta | null = useMemo(() => {
+    if (selectedStep === 'all') return null
+    const group = stepGroupsByName.get(selectedStep)
+    if (!group) return null
+    return {
+      command: group.command,
+      status: group.status,
+      durationMs: group.durationMs,
+    }
+  }, [selectedStep, stepGroupsByName])
+
+  const virtualizer = useVirtualizer({
+    count: filteredLogs.length,
+    getScrollElement: () => scrollContainerRef.current,
+    estimateSize: () => 20,
+    overscan: 50,
+  })
+  useAutoScroll(virtualizer, filteredLogs.length, autoScroll)
+
+  const handleScroll = useCallback(() => {
+    const element = scrollContainerRef.current
+    if (!element) return
+    setAutoScroll(
+      element.scrollHeight - element.scrollTop - element.clientHeight < 40,
+    )
+  }, [])
+
+  useMountEffect(() => {
+    const element = scrollContainerRef.current
+    if (!element) return
+    element.addEventListener('scroll', handleScroll)
+    return () => element.removeEventListener('scroll', handleScroll)
+  })
+
+  const openSearch = useCallback(() => {
+    setSearchOpen(true)
+    setTimeout(() => searchInputRef.current?.focus(), 0)
+  }, [])
+  const closeSearch = useCallback(() => {
+    setSearchOpen(false)
+    setSearchQuery('')
+  }, [])
+
+  useWindowEvent('keydown', (event) => {
+    if ((event.ctrlKey || event.metaKey) && event.key === 'f') {
+      const target = event.target as HTMLElement | null
+      if (target?.closest('input, textarea, [contenteditable="true"]')) return
+      const element = scrollContainerRef.current
+      if (!element) return
+      const rect = element.getBoundingClientRect()
+      if (rect.top < window.innerHeight && rect.bottom > 0) {
+        event.preventDefault()
+        openSearch()
+      }
+    }
+    if (event.key === 'Escape' && searchOpen) closeSearch()
+  })
+
+  const logStepGroups = stepGroups.filter((group) => group.logs.length > 0)
+  const hasSteps = logStepGroups.length > 0
+
+  function jumpToFirstError() {
+    const index = findFirstErrorIndex(filteredLogs)
+    if (index < 0) return
+    setAutoScroll(false)
+    virtualizer.scrollToIndex(index, { align: 'center' })
+  }
+
+  function downloadRawLogs() {
+    const blob = new Blob(
+      [selectedLogs.map((chunk) => chunk.content).join('\n')],
+      {
+        type: 'text/plain',
+      },
+    )
+    const url = URL.createObjectURL(blob)
+    const anchor = document.createElement('a')
+    anchor.href = url
+    anchor.download = 'build-logs.txt'
+    anchor.click()
+    URL.revokeObjectURL(url)
+  }
+
+  function scrollToLatest() {
+    setAutoScroll(true)
+    virtualizer.scrollToIndex(filteredLogs.length - 1, { align: 'end' })
+  }
+
+  return (
+    <div className="flex h-[60dvh] min-h-96 max-h-[48rem] flex-col">
+      <LogToolbar
+        isStreaming={isStreaming}
+        logCount={filteredLogs.length}
+        searchOpen={searchOpen}
+        searchQuery={searchQuery}
+        searchInputRef={searchInputRef}
+        wrapLines={wrapLines}
+        showScrollLatest={!autoScroll && filteredLogs.length > 0}
+        onSearchOpen={openSearch}
+        onSearchClose={closeSearch}
+        onSearchQueryChange={setSearchQuery}
+        onJumpToError={jumpToFirstError}
+        onToggleWrap={() => setWrapLines((value) => !value)}
+        onDownload={downloadRawLogs}
+        onScrollLatest={scrollToLatest}
+      />
+
+      {streamError ? (
+        <Alert variant="destructive" className="rounded-none border-b-0 py-2">
+          <HugeiconsIcon icon={AlertCircleIcon} />
+          <AlertDescription>{streamError}</AlertDescription>
+        </Alert>
+      ) : null}
+
+      {hasSteps ? (
+        <MobileStepSelector
+          groups={logStepGroups}
+          selectedStep={selectedStep}
+          allLogCount={allVisibleLogs.length}
+          onSelect={setUserSelectedStep}
+        />
+      ) : null}
+
+      <div
+        className={
+          hasSteps
+            ? 'grid min-h-0 flex-1 grid-cols-1 overflow-hidden border md:grid-cols-[220px_minmax(0,1fr)]'
+            : 'min-h-0 flex-1 overflow-hidden border'
+        }
+      >
+        {hasSteps ? (
+          <StepSidebar
+            groups={logStepGroups}
+            selectedStep={selectedStep}
+            allLogCount={allVisibleLogs.length}
+            onSelect={setUserSelectedStep}
+          />
+        ) : null}
+        <LogOutput
+          logs={filteredLogs}
+          selectedStep={selectedStep}
+          selectedStepMeta={selectedStepMeta}
+          searchQuery={searchQuery}
+          logsUnavailable={logsUnavailable}
+          isTerminal={isTerminal}
+          wrapLines={wrapLines}
+          scrollContainerRef={scrollContainerRef}
+          virtualizer={virtualizer}
+        />
+      </div>
+    </div>
+  )
+}
